requirements.txt
google-genai>=1.0.0
yfinance>=0.2.36
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
matplotlib>=3.7.0
moviepy==1.0.3
imageio[ffmpeg]
imageio-ffmpeg
gTTS>=2.3.2
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.9.0
